requirements.txt
python-telegram-bot==20.6
playwright==1.44.0
